pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Software Engineering course project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
jinja2 = "^3.1.4"
pytest = "^8.3.3"
httpx = "^0.27.2"
pypika = "^0.48.9"
contractspy = "^1.0.7"
passlib = "^1.7.4"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
generate-fastapi = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
